refactor(businesses): restructure businesses page into tabbed layout
Split the businesses page into separate tab components for Onboarded and Onboarding businesses to improve maintainability and readability. Removed redundant code and consolidated shared logic into reusable components.
